refactor: change form style. Move authorization file to folder. Change static AuthService to abstract class
