+++ src/views/authorization/logoutView.ts
@@ -1,6 +1,6 @@
-import RouterController from '../controllers/routerController';
-import UserController from '../controllers/userController';
-import ViewInterface from './viewInterface';
+import RouterController from '../../controllers/routerController';
+import UserController from '../../controllers/userController';
+import ViewInterface from '../viewInterface';
 
 class LogoutView extends ViewInterface {
     controller: UserController;
