+++ src/views/authorization/signinView.ts
@@ -1,9 +1,8 @@
-import UserController from '../controllers/userController';
-import { assertDefined } from '../helpers/helpers';
-import FormService from '../helpers/formService';
-import ViewInterface from './viewInterface';
+import UserController from '../../controllers/userController';
+import { assertDefined } from '../../helpers/helpers';
+import Authorization from './authorization';
 import './authView.css';
-class SigninView extends ViewInterface {
+class SigninView extends Authorization {
     controller: UserController;
     constructor(rootElement: HTMLElement) {
         super(rootElement);
@@ -11,7 +10,7 @@ class SigninView extends ViewInterface {
     }
 
     show(): void {
-        const form = FormService.getAuthForm();
+        const form = this.getAuthForm();
         form.classList.add('signin-form');
         this.fillLoginForm(form);
         form.addEventListener('input', () =>
@@ -25,14 +24,14 @@ class SigninView extends ViewInterface {
     }
 
     private fillLoginForm(form: HTMLElement) {
-        const email = FormService.getInput(['login-form__login'], 'email', 'Email...', 'email');
+        const email = this.getInput(['login-form__login'], 'email', 'Email...', 'email');
         form.append(email);
-        const pass = FormService.getInput(['login-form__pass'], 'password', 'Пароль...', 'password');
+        const pass = this.getInput(['login-form__pass'], 'password', 'Пароль...', 'password');
         email.required = true;
         form.append(pass);
         pass.minLength = 8;
         pass.required = true;
-        const submit = FormService.getSubmitBtn(['login-form__submit'], 'Войти', this.signinAction);
+        const submit = this.getSubmitBtn(['login-form__submit'], 'Войти', this.signinAction);
         form.append(submit);
     }
     private signinAction() {
